Added tests for QuestionService

+++ src/main/java/com/idatt2105/backend/service/QuestionService.java
@@ -2,6 +2,7 @@ package com.idatt2105.backend.service;
 
 import java.util.List;
 
+import java.util.Objects;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
@@ -86,7 +87,11 @@ public class QuestionService {
    * @return (Question) The question with the given id.
    * @throws InvalidIdException if the question with the given id is not found.
    */
-  public Question getQuestionById(@NotNull Long id) {
+  public Question getQuestionById(Long id) {
+    if (id == null) {
+      throw new IllegalArgumentException("Id parameter cannot be null.");
+    }
+
     return questionRepository
         .findById(id)
         .orElseThrow(() -> new InvalidIdException("Question with id " + id + " not found"));
@@ -98,7 +103,11 @@ public class QuestionService {
    * @param id (Long) The id of the question to delete.
    * @throws InvalidIdException if the question with the given id is not found.
    */
-  public void deleteQuestion(@NotNull Long id) {
+  public void deleteQuestion(Long id) {
+    if (id == null) {
+      throw new IllegalArgumentException("Id parameter cannot be null.");
+    }
+
     Question question = getQuestionById(id);
     questionRepository.delete(question);
   }
@@ -110,7 +119,11 @@ public class QuestionService {
    * @return (Question) The updated question.
    * @throws InvalidIdException if the question with the given id is not found.
    */
-  public Question updateQuestion(@Validated @NotNull QuestionDTO questionDTO) {
+  public Question updateQuestion(QuestionDTO questionDTO) {
+    if (questionDTO == null) {
+      throw new IllegalArgumentException("Question parameter cannot be null.");
+    }
+
     Question question = getQuestionById(questionDTO.getQuestionId());
     question.setQuestionText(questionDTO.getQuestionText());
     question.setMediaUrl(questionDTO.getMediaUrl());
@@ -125,8 +138,14 @@ public class QuestionService {
    * @param questionDTO (QuestionDTO) Data transfer object for the question.
    * @return (TrueOrFalseQuestion) The updated true or false question.
    */
-  public TrueOrFalseQuestion updateTrueOrFalseQuestion(
-      @Validated @NotNull QuestionDTO questionDTO) {
+  public TrueOrFalseQuestion updateTrueOrFalseQuestion(QuestionDTO questionDTO) {
+    if (questionDTO == null) {
+      throw new IllegalArgumentException("Question parameter cannot be null.");
+    }
+    if (questionDTO.isCorrect() == null) {
+      throw new IllegalArgumentException("IsCorrect field must have a value.");
+    }
+
     Question question = getQuestionById(questionDTO.getQuestionId());
     TrueOrFalseQuestion trueOrFalseQuestion;
     try {
@@ -145,7 +164,10 @@ public class QuestionService {
    * @param quizId (Long) The id of the quiz to get questions from.
    * @return (List&lt;Question&gt;) All questions in the quiz with the given id.
    */
-  public List<Question> getQuestionsByQuizId(@NotNull Long quizId) {
+  public List<Question> getQuestionsByQuizId(Long quizId) {
+    if (quizId == null) {
+      throw new IllegalArgumentException("Quiz id parameter cannot be null.");
+    }
     return questionRepository.findQuestionsByQuizId(quizId);
   }
 
@@ -156,7 +178,10 @@ public class QuestionService {
    * @throws InvalidIdException if the question with the given id is not found or is not a multiple
    *     choice question.
    */
-  public Alternative addAlternative(@Validated @NotNull AlternativeDTO alternativeDTO) {
+  public Alternative addAlternative(AlternativeDTO alternativeDTO) {
+    if (alternativeDTO == null) {
+      throw new IllegalArgumentException("Alternative parameter cannot be null.");
+    }
     Question q = getQuestionById(alternativeDTO.getQuestionId());
     MultipleChoiceQuestion question;
     try {
@@ -177,27 +202,34 @@ public class QuestionService {
    *
    * @param id (Long) The id of the alternative to delete.
    */
-  public void deleteAlternative(@Validated @NotNull Long id) {
+  public void deleteAlternative(Long id) {
+    if (id == null) {
+      throw new IllegalArgumentException("Id parameter cannot be null.");
+    }
     alternativeRepository.deleteById(id);
   }
 
-  public Question addTags(@Validated @NotNull QuestionDTO dto) {
-    Question question = getQuestionById(dto.getQuestionId());
+  public Question addTags(QuestionDTO dto) {
+    if (dto == null) {
+      throw new IllegalArgumentException("Question parameter cannot be null.");
+    }
 
-    // Verify that all tags exist
-    List<Long> allIds = dto.getAllTagIds();
-    List<Tag> existingTags = tagRepository.findAllById(allIds);
-    allIds.forEach(
-        id -> {
-          if (existingTags.stream().noneMatch(tag -> tag.getId().equals(id))) {
-            throw new InvalidIdException("Tag with id " + id + " not found");
-          }
-        });
+    Question question = getQuestionById(dto.getQuestionId());
+    dto.getTags().stream().filter(Objects::nonNull).forEach(tag -> {
+      tag.setId(null);
+      if (Boolean.FALSE.equals(tagRepository.existsByTagName(tag.getTagName()))) {
+        tagRepository.save(tag);
+      }
+    });
     question.addTags(dto.getTags());
     return questionRepository.save(question);
   }
 
-  public Question deleteTags(@Validated @NotNull QuestionDTO dto) {
+  public Question deleteTags(QuestionDTO dto) {
+    if (dto == null) {
+      throw new IllegalArgumentException("Question parameter cannot be null.");
+    }
+
     Question question = getQuestionById(dto.getQuestionId());
     question.removeTags(dto.getTags());
     return questionRepository.save(question);
